--- packages/proximal-testing-videos/src/no_audio_animation/global.css
@@ -1,1 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto:wght@400;700&display=swap');
