--- packages/proximal-testing-videos/src/no_audio_animation/project.ts
@@ -1,34 +1,0 @@
-import {Color, makeProject, Vector2} from '@revideo/core';
-
-import {scene} from './example';
-
-import './global.css';
-
-export const project = makeProject({
-  name: 'project',
-  scenes: [scene],
-  variables: {
-    fill: 'green',
-  },
-  settings: {
-    shared: {
-      background: '#FFFFFF',
-      range: [0, Infinity],
-      size: new Vector2(640, 480),
-    },
-    preview: {
-      fps: 30,
-      resolutionScale: 1,
-    },
-    rendering: {
-      exporter: {
-        name: '@revideo/core/wasm',
-      },
-      fps: 30,
-      resolutionScale: 1,
-      colorSpace: 'srgb',
-    },
-  },
-});
-
-export default project;
